resite/requirements.txt
fastapi
sqlalchemy
pydantic
aiomysql
pymysql
dotenv
bcrypt
passlib
python-jose